# perf: remove environment variable reading from init