This PR removes runtime environment variable read for DenseNDarray creation

- Before this PR
            ```
            %timeit aux = jina.types.ndarray.dense.numpy.DenseNdArray()
            2.81 µs ± 177 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
            ```
            
- After this PR
            ```
            %timeit aux = jina.types.ndarray.dense.numpy.DenseNdArray()
            1.48 µs ± 11.8 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
            ```